refactor(otp): extract reusable TOTP generator into runtime/otp
Move the RFC 6238 time-based one-time-password generator out of the mfa
generator and into a shared runtime/otp package, exporting GenerateCode
so providers that expose TOTP can reuse it rather than duplicating the
logic.

The mfa generator now imports runtime/otp; its behavior is unchanged and
its known-answer tests move with the code. Also drops a dead
`length > math.MaxInt` guard that could never fire (length is an int).
